+++ addOns/pscan/src/main/java/org/zaproxy/addon/pscan/PassiveScanApi.java
@@ -19,7 +19,6 @@
  */
 package org.zaproxy.addon.pscan;
 
-import java.lang.reflect.Method;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -28,7 +27,8 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.parosproxy.paros.Constant;
 import org.parosproxy.paros.core.scanner.Plugin;
-import org.zaproxy.addon.pscan.internal.ScanRuleManager;
+import org.zaproxy.addon.pscan.internal.PassiveScannerOptions;
+import org.zaproxy.addon.pscan.internal.scanner.PassiveScanTask;
 import org.zaproxy.zap.extension.api.ApiAction;
 import org.zaproxy.zap.extension.api.ApiException;
 import org.zaproxy.zap.extension.api.ApiImplementor;
@@ -37,9 +37,6 @@ import org.zaproxy.zap.extension.api.ApiResponseElement;
 import org.zaproxy.zap.extension.api.ApiResponseList;
 import org.zaproxy.zap.extension.api.ApiResponseSet;
 import org.zaproxy.zap.extension.api.ApiView;
-import org.zaproxy.zap.extension.pscan.ExtensionPassiveScan;
-import org.zaproxy.zap.extension.pscan.PassiveScanParam;
-import org.zaproxy.zap.extension.pscan.PassiveScanTask;
 import org.zaproxy.zap.extension.pscan.PassiveScanner;
 import org.zaproxy.zap.extension.pscan.PluginPassiveScanner;
 import org.zaproxy.zap.utils.ApiUtils;
@@ -76,17 +73,16 @@ public class PassiveScanApi extends ApiImplementor {
     private static final String PARAM_ALERT_THRESHOLD = "alertThreshold";
     private static final String PARAM_MAX_ALERTS = "maxAlerts";
 
-    private ExtensionPassiveScan extension;
-    private final ScanRuleManager scanRuleManager;
-    private Method setPassiveScanEnabledMethod;
+    private final ExtensionPassiveScan2 extension;
+    private final PassiveScannersManager scannersManager;
 
     public PassiveScanApi() {
         this(null, null);
     }
 
-    public PassiveScanApi(ExtensionPassiveScan extension, ScanRuleManager scanRuleManager) {
+    public PassiveScanApi(ExtensionPassiveScan2 extension, PassiveScannersManager scannersManager) {
         this.extension = extension;
-        this.scanRuleManager = scanRuleManager;
+        this.scannersManager = scannersManager;
 
         this.addApiAction(new ApiAction(ACTION_SET_ENABLED, new String[] {PARAM_ENABLED}));
         this.addApiAction(
@@ -110,23 +106,11 @@ public class PassiveScanApi extends ApiImplementor {
         this.addApiView(new ApiView(VIEW_SCANNERS));
         ApiView currentRule = new ApiView(VIEW_CURRENT_RULE);
         currentRule.setDeprecated(true);
-        // When generating the API clients messages is not initialised.
-        if (Constant.messages != null) {
-            currentRule.setDeprecatedDescription(
-                    Constant.messages.getString("pscan.api.view.currentRule.deprecated"));
-        }
+        currentRule.setDeprecatedDescription(
+                Constant.messages.getString("pscan.api.view.currentRule.deprecated"));
         this.addApiView(currentRule);
         this.addApiView(new ApiView(VIEW_CURRENT_TASKS));
         this.addApiView(new ApiView(VIEW_MAX_ALERTS_PER_RULE));
-
-        try {
-            setPassiveScanEnabledMethod =
-                    ExtensionPassiveScan.class.getDeclaredMethod(
-                            "setPassiveScanEnabled", boolean.class);
-            setPassiveScanEnabledMethod.setAccessible(true);
-        } catch (NoSuchMethodException | SecurityException e) {
-            LOGGER.debug("An error occurred while getting the method:", e);
-        }
     }
 
     @Override
@@ -140,14 +124,10 @@ public class PassiveScanApi extends ApiImplementor {
             case ACTION_SET_ENABLED:
                 boolean enabled = getParam(params, PARAM_ENABLED, false);
 
-                try {
-                    setPassiveScanEnabledMethod.invoke(extension, enabled);
-                } catch (Exception e) {
-                    throw new ApiException(ApiException.Type.INTERNAL_ERROR, e);
-                }
+                extension.setPassiveScanEnabled(enabled);
                 break;
             case ACTION_SET_SCAN_ONLY_IN_SCOPE:
-                getPassiveScanParam().setScanOnlyInScope(params.getBoolean(PARAM_ONLY_IN_SCOPE));
+                getOptions().setScanOnlyInScope(params.getBoolean(PARAM_ONLY_IN_SCOPE));
                 break;
             case ACTION_ENABLE_ALL_SCANNERS:
                 setAllPluginPassiveScannersEnabled(true);
@@ -178,16 +158,15 @@ public class PassiveScanApi extends ApiImplementor {
                 setPluginPassiveScannerAlertThreshold(pluginId, alertThreshold);
                 break;
             case ACTION_SET_MAX_ALERTS_PER_RULE:
-                getPassiveScanParam()
-                        .setMaxAlertsPerRule(ApiUtils.getIntParam(params, PARAM_MAX_ALERTS));
+                getOptions().setMaxAlertsPerRule(ApiUtils.getIntParam(params, PARAM_MAX_ALERTS));
                 break;
             case ACTION_DISABLE_ALL_TAGS:
-                getPassiveScanParam()
+                getOptions()
                         .getAutoTagScanners()
                         .forEach(tagScanner -> tagScanner.setEnabled(false));
                 break;
             case ACTION_ENABLE_ALL_TAGS:
-                getPassiveScanParam()
+                getOptions()
                         .getAutoTagScanners()
                         .forEach(tagScanner -> tagScanner.setEnabled(true));
                 break;
@@ -201,8 +180,8 @@ public class PassiveScanApi extends ApiImplementor {
         return ApiResponseElement.OK;
     }
 
-    private PassiveScanParam getPassiveScanParam() {
-        return extension.getModel().getOptionsParam().getParamSet(PassiveScanParam.class);
+    private PassiveScannerOptions getOptions() {
+        return extension.getModel().getOptionsParam().getParamSet(PassiveScannerOptions.class);
     }
 
     private void setPluginPassiveScannersEnabled(JSONObject params, boolean enabled)
@@ -233,10 +212,7 @@ public class PassiveScanApi extends ApiImplementor {
     }
 
     private PluginPassiveScanner getScanRule(int pluginId) {
-        if (scanRuleManager != null) {
-            return (PluginPassiveScanner) scanRuleManager.getScanRule(pluginId);
-        }
-        return extension.getPluginPassiveScanner(pluginId);
+        return scannersManager.getScanRule(pluginId);
     }
 
     /**
@@ -252,10 +228,7 @@ public class PassiveScanApi extends ApiImplementor {
     }
 
     private List<PluginPassiveScanner> getPluginScanRules() {
-        if (scanRuleManager != null) {
-            return scanRuleManager.getPluginScanRules();
-        }
-        return extension.getPluginPassiveScanners();
+        return scannersManager.getScanRules();
     }
 
     /**
@@ -313,7 +286,7 @@ public class PassiveScanApi extends ApiImplementor {
             case VIEW_SCAN_ONLY_IN_SCOPE:
                 result =
                         new ApiResponseElement(
-                                name, Boolean.toString(getPassiveScanParam().isScanOnlyInScope()));
+                                name, Boolean.toString(getOptions().isScanOnlyInScope()));
                 break;
             case VIEW_RECORDS_TO_SCAN:
                 result = new ApiResponseElement(name, String.valueOf(extension.getRecordsToScan()));
@@ -348,7 +321,7 @@ public class PassiveScanApi extends ApiImplementor {
                 result =
                         new ApiResponseElement(
                                 VIEW_MAX_ALERTS_PER_RULE,
-                                Integer.toString(getPassiveScanParam().getMaxAlertsPerRule()));
+                                Integer.toString(getOptions().getMaxAlertsPerRule()));
                 break;
             default:
                 throw new ApiException(ApiException.Type.BAD_VIEW);
@@ -358,14 +331,16 @@ public class PassiveScanApi extends ApiImplementor {
 
     private ApiResponseSet<String> getResponseForTask(PassiveScanTask task, String name) {
         Map<String, String> map = new HashMap<>();
-        PassiveScanner scanner = task.getCurrentScanner();
-        map.put("name", scanner == null ? "" : scanner.getName());
-        map.put("url", task.getURI().toString());
-        long time = task.getStartTime();
-        if (time > 0) {
-            time = System.currentTimeMillis() - time;
+        if (task != null) {
+            PassiveScanner scanner = task.getCurrentScanner();
+            map.put("name", scanner == null ? "" : scanner.getName());
+            map.put("url", task.getURI().toString());
+            long time = task.getStartTime();
+            if (time > 0) {
+                time = System.currentTimeMillis() - time;
+            }
+            map.put("time", String.valueOf(time));
         }
-        map.put("time", String.valueOf(time));
         return new ApiResponseSet<>(name, map);
     }
 }